btc from double -> BigDecimal

+++ core/src/main/java/io/github/felipebueno/core/ExchangeRate.java
@@ -1,13 +1,15 @@
 package io.github.felipebueno.core;
 
+import java.math.BigDecimal;
+
 public class ExchangeRate {
 
 	public final String code;
 	public final double rate;
-	public final double balance;
+	public final BigDecimal balance;
 	public boolean isDefault;
 
-	public ExchangeRate(String code, double rate, double balance, boolean isDefault) {
+	public ExchangeRate(String code, double rate, BigDecimal balance, boolean isDefault) {
 		this.code = code;
 		this.rate = rate;
 		this.balance = balance;
